Agent: Move connect_to_dc to vuln_assessment.py
