+++ monkey/infection_monkey/exploit/zerologon.py
@@ -12,11 +12,10 @@ from binascii import unhexlify
 from typing import Dict, List, Optional, Sequence, Tuple
 
 import impacket
-from impacket.dcerpc.v5 import epm, nrpc, rpcrt, transport
+from impacket.dcerpc.v5 import nrpc, rpcrt
 from impacket.dcerpc.v5.dtypes import NULL
 
 from common.agent_events import CredentialsStolenEvent, PasswordRestorationEvent
-from common.common_consts.timeouts import LONG_REQUEST_TIMEOUT
 from common.credentials import Credentials, LMHash, NTHash, Username
 from common.tags import (
     T1003_ATTACK_TECHNIQUE_TAG,
@@ -27,7 +26,11 @@ from infection_monkey.exploit.HostExploiter import HostExploiter
 from infection_monkey.exploit.tools.wmi_tools import WmiTools
 from infection_monkey.exploit.zerologon_utils.dump_secrets import DumpSecrets
 from infection_monkey.exploit.zerologon_utils.options import OptionsForSecretsdump
-from infection_monkey.exploit.zerologon_utils.vuln_assessment import get_dc_details, is_exploitable
+from infection_monkey.exploit.zerologon_utils.vuln_assessment import (
+    connect_to_dc,
+    get_dc_details,
+    is_exploitable,
+)
 from infection_monkey.exploit.zerologon_utils.wmiexec import Wmiexec
 from infection_monkey.i_puppet import ExploiterResultData
 from infection_monkey.utils.capture_output import StdoutCapture
@@ -112,16 +115,6 @@ class ZerologonExploiter(HostExploiter):
             logger.info("System was not exploited.")
 
         return self.exploit_result
-
-    @staticmethod
-    def connect_to_dc(dc_ip) -> object:
-        binding = epm.hept_map(dc_ip, nrpc.MSRPC_UUID_NRPC, protocol="ncacn_ip_tcp")
-        rpc_transport = transport.DCERPCTransportFactory(binding)
-        rpc_transport.set_connect_timeout(LONG_REQUEST_TIMEOUT)
-        rpc_con = rpc_transport.get_dce_rpc()
-        rpc_con.connect()
-        rpc_con.bind(nrpc.MSRPC_UUID_NRPC)
-        return rpc_con
 
     def _send_exploit_rpc_login_requests(self, rpc_con) -> bool:
         for _ in interruptible_iter(range(0, self.MAX_ATTEMPTS), self.interrupt):
@@ -230,7 +223,7 @@ class ZerologonExploiter(HostExploiter):
 
             # Connect to the DC's Netlogon service.
             try:
-                rpc_con = ZerologonExploiter.connect_to_dc(self.dc_ip)
+                rpc_con = connect_to_dc(self.dc_ip)
             except Exception as e:
                 logger.info(f"Exception occurred while connecting to DC: {str(e)}")
                 return False
