+++ monkey/infection_monkey/exploit/zerologon_utils/vuln_assessment.py
@@ -2,14 +2,24 @@ import logging
 from typing import Optional, Tuple
 
 import nmb.NetBIOS
-from impacket.dcerpc.v5 import nrpc, rpcrt
+from impacket.dcerpc.v5 import epm, nrpc, rpcrt, transport
 
-from common.common_consts.timeouts import MEDIUM_REQUEST_TIMEOUT
+from common.common_consts.timeouts import LONG_REQUEST_TIMEOUT, MEDIUM_REQUEST_TIMEOUT
 from common.utils.exceptions import DomainControllerNameFetchError
 from infection_monkey.model import VictimHost
 from infection_monkey.utils.threading import interruptible_iter
 
 logger = logging.getLogger(__name__)
+
+
+def connect_to_dc(dc_ip) -> object:
+    binding = epm.hept_map(dc_ip, nrpc.MSRPC_UUID_NRPC, protocol="ncacn_ip_tcp")
+    rpc_transport = transport.DCERPCTransportFactory(binding)
+    rpc_transport.set_connect_timeout(LONG_REQUEST_TIMEOUT)
+    rpc_con = rpc_transport.get_dce_rpc()
+    rpc_con.connect()
+    rpc_con.bind(nrpc.MSRPC_UUID_NRPC)
+    return rpc_con
 
 
 def get_dc_details(host: VictimHost) -> Tuple[str, str, str]:
@@ -39,7 +49,7 @@ def _get_dc_name(dc_ip: str) -> str:
 def is_exploitable(zerologon_exploiter_object) -> Tuple[bool, Optional[rpcrt.DCERPC_v5]]:
     # Connect to the DC's Netlogon service.
     try:
-        rpc_con = zerologon_exploiter_object.connect_to_dc(zerologon_exploiter_object.dc_ip)
+        rpc_con = connect_to_dc(zerologon_exploiter_object.dc_ip)
     except Exception as err:
         error_message = f"Exception occurred while connecting to DC: {err}"
         logger.info(error_message)
